HackPrinceton/backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
psycopg[binary]==3.2.12
python-dotenv==1.0.1
stripe==11.5.0
requests==2.32.3
python-jose[cryptography]==3.3.0